fix referential_integrity for postgres partitioned tables
In Postgres, calling `ActiveRecord::ConnectionAdapters::PostgreSQL::ReferentialIntegrity#all_foreign_keys_valid!` with partioned tables leaves some partition in NOT VALID state.
This is because partitions can have foreign key constraints with the same name as partitioned table. The `UDPATE` is not scoped to a specific table so it will mark all partitions as invalid, while only validating one table, leaving all the others invalid. By scoping the `UPDATE` to a specific table we ensure that we touch only the partition we are about to validate.